Refactor: Modularize time_tools into a dedicated package with schemas
