+++ agent/core_agent/tools/time_tools/README.md
@@ -1,0 +1,3 @@
+# Time Tools
+
+This tool provides the agent with the current system time, enabling time-aware reasoning.
